+++ src/components/Feedback/Feedback.js
@@ -1,32 +1,35 @@
 /**
  * React Starter Kit (https://www.reactstarterkit.com/)
  *
- * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
+ * Copyright © 2014-present Kriasoft, LLC. All rights reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE.txt file in the root directory of this source tree.
  */
 
+import useStyles from 'isomorphic-style-loader/useStyles';
 import React from 'react';
-import withStyles from 'isomorphic-style-loader/lib/withStyles';
 import s from './Feedback.css';
 
-function Feedback() {
+export default function Feedback() {
+  useStyles(s);
   return (
     <div className={s.root}>
       <div className={s.container}>
         <a
           className={s.link}
           href="https://gitter.im/kriasoft/react-starter-kit"
-        >Ask a question</a>
+        >
+          Ask a question
+        </a>
         <span className={s.spacer}>|</span>
         <a
           className={s.link}
           href="https://github.com/kriasoft/react-starter-kit/issues/new"
-        >Report an issue</a>
+        >
+          Report an issue
+        </a>
       </div>
     </div>
   );
 }
-
-export default withStyles(s)(Feedback);