+++ src/components/Header/Header.js
@@ -1,26 +1,34 @@
 /**
  * React Starter Kit (https://www.reactstarterkit.com/)
  *
- * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
+ * Copyright © 2014-present Kriasoft, LLC. All rights reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE.txt file in the root directory of this source tree.
  */
 
+import useStyles from 'isomorphic-style-loader/useStyles';
 import React from 'react';
-import withStyles from 'isomorphic-style-loader/lib/withStyles';
 import s from './Header.css';
 import Link from '../Link';
 import Navigation from '../Navigation';
 import logoUrl from './logo-small.png';
+import logoUrl2x from './logo-small@2x.png';
 
-function Header() {
+export default function Header() {
+  useStyles(s);
   return (
     <div className={s.root}>
       <div className={s.container}>
-        <Navigation className={s.nav} />
+        <Navigation />
         <Link className={s.brand} to="/">
-          <img src={logoUrl} width="38" height="38" alt="React" />
+          <img
+            src={logoUrl}
+            srcSet={`${logoUrl2x} 2x`}
+            width="38"
+            height="38"
+            alt="React"
+          />
           <span className={s.brandTxt}>Your Company</span>
         </Link>
         <div className={s.banner}>
@@ -31,5 +39,3 @@ function Header() {
     </div>
   );
 }
-
-export default withStyles(s)(Header);