+++ src/data/models/User.js
@@ -1,7 +1,7 @@
 /**
  * React Starter Kit (https://www.reactstarterkit.com/)
  *
- * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
+ * Copyright © 2014-present Kriasoft, LLC. All rights reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE.txt file in the root directory of this source tree.
@@ -10,30 +10,28 @@
 import DataType from 'sequelize';
 import Model from '../sequelize';
 
-const User = Model.define('User', {
+const User = Model.define(
+  'User',
+  {
+    id: {
+      type: DataType.UUID,
+      defaultValue: DataType.UUIDV1,
+      primaryKey: true,
+    },
 
-  id: {
-    type: DataType.UUID,
-    defaultValue: DataType.UUIDV1,
-    primaryKey: true,
+    email: {
+      type: DataType.STRING(255),
+      validate: { isEmail: true },
+    },
+
+    emailConfirmed: {
+      type: DataType.BOOLEAN,
+      defaultValue: false,
+    },
   },
-
-  email: {
-    type: DataType.STRING(255),
-    validate: { isEmail: true },
+  {
+    indexes: [{ fields: ['email'] }],
   },
-
-  emailConfirmed: {
-    type: DataType.BOOLEAN,
-    defaultValue: false,
-  },
-
-}, {
-
-  indexes: [
-    { fields: ['email'] },
-  ],
-
-});
+);
 
 export default User;