+++ src/data/models/UserClaim.js
@@ -1,7 +1,7 @@
 /**
  * React Starter Kit (https://www.reactstarterkit.com/)
  *
- * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
+ * Copyright © 2014-present Kriasoft, LLC. All rights reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE.txt file in the root directory of this source tree.
@@ -11,7 +11,6 @@ import DataType from 'sequelize';
 import Model from '../sequelize';
 
 const UserClaim = Model.define('UserClaim', {
-
   type: {
     type: DataType.STRING,
   },
@@ -19,7 +18,6 @@ const UserClaim = Model.define('UserClaim', {
   value: {
     type: DataType.STRING,
   },
-
 });
 
 export default UserClaim;