+++ src/data/models/UserLogin.js
@@ -1,7 +1,7 @@
 /**
  * React Starter Kit (https://www.reactstarterkit.com/)
  *
- * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
+ * Copyright © 2014-present Kriasoft, LLC. All rights reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE.txt file in the root directory of this source tree.
@@ -11,7 +11,6 @@ import DataType from 'sequelize';
 import Model from '../sequelize';
 
 const UserLogin = Model.define('UserLogin', {
-
   name: {
     type: DataType.STRING(50),
     primaryKey: true,
@@ -21,7 +20,6 @@ const UserLogin = Model.define('UserLogin', {
     type: DataType.STRING(100),
     primaryKey: true,
   },
-
 });
 
 export default UserLogin;