+++ src/routes/error/ErrorPage.js
@@ -1,40 +1,48 @@
 /**
  * React Starter Kit (https://www.reactstarterkit.com/)
  *
- * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
+ * Copyright © 2014-present Kriasoft, LLC. All rights reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE.txt file in the root directory of this source tree.
  */
 
-import React, { PropTypes } from 'react';
-import withStyles from 'isomorphic-style-loader/lib/withStyles';
+import useStyles from 'isomorphic-style-loader/useStyles';
+import React from 'react';
+import PropTypes from 'prop-types';
 import s from './ErrorPage.css';
 
-function ErrorPage({ error }, context) {
-  let title = 'Error';
-  let content = 'Sorry, a critical error occurred on this page.';
-  let errorMessage = null;
-
-  if (error.status === 404) {
-    title = 'Page Not Found';
-    content = 'Sorry, the page you were trying to view does not exist.';
-  } else if (process.env.NODE_ENV !== 'production') {
-    errorMessage = <pre>{error.stack}</pre>;
+export function ErrorPageWithoutStyle({ error }) {
+  if (__DEV__ && error) {
+    return (
+      <>
+        <h1>{error.name}</h1>
+        <pre>{error.stack}</pre>
+      </>
+    );
   }
 
-  context.setTitle(title);
-
   return (
-    <div>
-      <h1>{title}</h1>
-      <p>{content}</p>
-      {errorMessage}
-    </div>
+    <>
+      <h1>Error</h1>
+      <p>Sorry, a critical error occurred on this page.</p>
+    </>
   );
 }
 
-ErrorPage.propTypes = { error: PropTypes.object.isRequired };
-ErrorPage.contextTypes = { setTitle: PropTypes.func.isRequired };
+ErrorPageWithoutStyle.propTypes = {
+  error: PropTypes.shape({
+    name: PropTypes.string.isRequired,
+    message: PropTypes.string.isRequired,
+    stack: PropTypes.string.isRequired,
+  }),
+};
 
-export default withStyles(s)(ErrorPage);
+ErrorPageWithoutStyle.defaultProps = {
+  error: null,
+};
+
+export default function ErrorPage(props) {
+  useStyles(s);
+  return ErrorPageWithoutStyle(props);
+}