+++ src/routes/login/index.js
@@ -1,21 +1,28 @@
 /**
  * React Starter Kit (https://www.reactstarterkit.com/)
  *
- * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
+ * Copyright © 2014-present Kriasoft, LLC. All rights reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE.txt file in the root directory of this source tree.
  */
 
 import React from 'react';
+import Layout from '../../components/Layout';
 import Login from './Login';
 
-export default {
+const title = 'Log In';
 
-  path: '/login',
+function action() {
+  return {
+    chunks: ['login'],
+    title,
+    component: (
+      <Layout>
+        <Login title={title} />
+      </Layout>
+    ),
+  };
+}
 
-  action() {
-    return <Login />;
-  },
-
-};
+export default action;