+++ src/routes/register/Register.js
@@ -1,20 +1,19 @@
 /**
  * React Starter Kit (https://www.reactstarterkit.com/)
  *
- * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
+ * Copyright © 2014-present Kriasoft, LLC. All rights reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE.txt file in the root directory of this source tree.
  */
 
-import React, { PropTypes } from 'react';
-import withStyles from 'isomorphic-style-loader/lib/withStyles';
+import useStyles from 'isomorphic-style-loader/useStyles';
+import React from 'react';
+import PropTypes from 'prop-types';
 import s from './Register.css';
 
-const title = 'New User Registration';
-
-function Register(props, context) {
-  context.setTitle(title);
+export default function Register({ title }) {
+  useStyles(s);
   return (
     <div className={s.root}>
       <div className={s.container}>
@@ -25,6 +24,6 @@ function Register(props, context) {
   );
 }
 
-Register.contextTypes = { setTitle: PropTypes.func.isRequired };
-
-export default withStyles(s)(Register);
+Register.propTypes = {
+  title: PropTypes.string.isRequired,
+};