+++ tools/webpack.config.js
@@ -1,33 +1,37 @@
 /**
  * React Starter Kit (https://www.reactstarterkit.com/)
  *
- * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
+ * Copyright © 2014-present Kriasoft, LLC. All rights reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE.txt file in the root directory of this source tree.
  */
 
+import fs from 'fs';
 import path from 'path';
 import webpack from 'webpack';
-import extend from 'extend';
-import AssetsPlugin from 'assets-webpack-plugin';
+import WebpackAssetsManifest from 'webpack-assets-manifest';
+import nodeExternals from 'webpack-node-externals';
+import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer';
+import overrideRules from './lib/overrideRules';
+import pkg from '../package.json';
 
-const DEBUG = !process.argv.includes('--release');
-const VERBOSE = process.argv.includes('--verbose');
-const AUTOPREFIXER_BROWSERS = [
-  'Android 2.3',
-  'Android >= 4',
-  'Chrome >= 35',
-  'Firefox >= 31',
-  'Explorer >= 9',
-  'iOS >= 7',
-  'Opera >= 12',
-  'Safari >= 7.1',
-];
-const GLOBALS = {
-  'process.env.NODE_ENV': DEBUG ? '"development"' : '"production"',
-  __DEV__: DEBUG,
-};
+const ROOT_DIR = path.resolve(__dirname, '..');
+const resolvePath = (...args) => path.resolve(ROOT_DIR, ...args);
+const SRC_DIR = resolvePath('src');
+const BUILD_DIR = resolvePath('build');
+
+const isDebug = !process.argv.includes('--release');
+const isVerbose = process.argv.includes('--verbose');
+const isAnalyze =
+  process.argv.includes('--analyze') || process.argv.includes('--analyse');
+
+const reScript = /\.(js|jsx|mjs)$/;
+const reStyle = /\.(css|less|styl|scss|sass|sss)$/;
+const reImage = /\.(bmp|gif|jpg|jpeg|png|svg)$/;
+const staticAssetName = isDebug
+  ? '[path][name].[ext]?[hash:8]'
+  : '[hash:8].[ext]';
 
 //
 // Common configuration chunk to be used for both
@@ -35,265 +39,445 @@ const GLOBALS = {
 // -----------------------------------------------------------------------------
 
 const config = {
-  context: path.resolve(__dirname, '../src'),
+  context: ROOT_DIR,
+
+  mode: isDebug ? 'development' : 'production',
 
   output: {
-    path: path.resolve(__dirname, '../build/public/assets'),
+    path: resolvePath(BUILD_DIR, 'public/assets'),
     publicPath: '/assets/',
-    sourcePrefix: '  ',
+    pathinfo: isVerbose,
+    filename: isDebug ? '[name].js' : '[name].[chunkhash:8].js',
+    chunkFilename: isDebug
+      ? '[name].chunk.js'
+      : '[name].[chunkhash:8].chunk.js',
+    // Point sourcemap entries to original disk location (format as URL on Windows)
+    devtoolModuleFilenameTemplate: info =>
+      path.resolve(info.absoluteResourcePath).replace(/\\/g, '/'),
+  },
+
+  resolve: {
+    // Allow absolute paths in imports, e.g. import Button from 'components/Button'
+    // Keep in sync with .flowconfig and .eslintrc
+    modules: ['node_modules', 'src'],
   },
 
   module: {
-    loaders: [
+    // Make missing exports an error instead of warning
+    strictExportPresence: true,
+
+    rules: [
+      // Rules for JS / JSX
       {
-        test: /\.jsx?$/,
+        test: reScript,
+        include: [SRC_DIR, resolvePath('tools')],
         loader: 'babel-loader',
-        include: [
-          path.resolve(__dirname, '../node_modules/react-routing/src'),
-          path.resolve(__dirname, '../src'),
-        ],
-        query: {
+        options: {
           // https://github.com/babel/babel-loader#options
-          cacheDirectory: DEBUG,
+          cacheDirectory: isDebug,
 
           // https://babeljs.io/docs/usage/options/
           babelrc: false,
+          configFile: false,
           presets: [
-            'react',
-            'es2015',
-            'stage-0',
+            // A Babel preset that can automatically determine the Babel plugins and polyfills
+            // https://github.com/babel/babel-preset-env
+            [
+              '@babel/preset-env',
+              {
+                targets: {
+                  browsers: pkg.browserslist,
+                },
+                forceAllTransforms: !isDebug, // for UglifyJS
+                modules: false,
+                useBuiltIns: false,
+                debug: false,
+              },
+            ],
+            // Flow
+            // https://github.com/babel/babel/tree/master/packages/babel-preset-flow
+            '@babel/preset-flow',
+            // JSX
+            // https://github.com/babel/babel/tree/master/packages/babel-preset-react
+            ['@babel/preset-react', { development: isDebug }],
           ],
           plugins: [
-            'transform-runtime',
-            ...DEBUG ? [] : [
-              'transform-react-remove-prop-types',
-              'transform-react-constant-elements',
-              'transform-react-inline-elements',
-            ],
+            // Experimental ECMAScript proposals
+            '@babel/plugin-proposal-class-properties',
+            '@babel/plugin-syntax-dynamic-import',
+            // Treat React JSX elements as value types and hoist them to the highest scope
+            // https://github.com/babel/babel/tree/master/packages/babel-plugin-transform-react-constant-elements
+            ...(isDebug ? [] : ['@babel/transform-react-constant-elements']),
+            // Replaces the React.createElement function with one that is more optimized for production
+            // https://github.com/babel/babel/tree/master/packages/babel-plugin-transform-react-inline-elements
+            ...(isDebug ? [] : ['@babel/transform-react-inline-elements']),
+            // Remove unnecessary React propTypes from the production build
+            // https://github.com/oliviertassinari/babel-plugin-transform-react-remove-prop-types
+            ...(isDebug ? [] : ['transform-react-remove-prop-types']),
           ],
         },
       },
+
+      // Rules for Style Sheets
       {
-        test: /\.css/,
-        loaders: [
-          'isomorphic-style-loader',
-          `css-loader?${JSON.stringify({
-            sourceMap: DEBUG,
-            // CSS Modules https://github.com/css-modules/css-modules
-            modules: true,
-            localIdentName: DEBUG ? '[name]_[local]_[hash:base64:3]' : '[hash:base64:4]',
-            // CSS Nano http://cssnano.co/options/
-            minimize: !DEBUG,
-          })}`,
-          'postcss-loader?pack=default',
+        test: reStyle,
+        rules: [
+          // Convert CSS into JS module
+          {
+            issuer: { not: [reStyle] },
+            use: 'isomorphic-style-loader',
+          },
+
+          // Process external/third-party styles
+          {
+            exclude: SRC_DIR,
+            loader: 'css-loader',
+            options: {
+              sourceMap: isDebug,
+            },
+          },
+
+          // Process internal/project styles (from src folder)
+          {
+            include: SRC_DIR,
+            loader: 'css-loader',
+            options: {
+              // CSS Loader https://github.com/webpack/css-loader
+              importLoaders: 1,
+              sourceMap: isDebug,
+              // CSS Modules https://github.com/css-modules/css-modules
+              modules: {
+                localIdentName: isDebug
+                  ? '[name]-[local]-[hash:base64:5]'
+                  : '[hash:base64:5]',
+              },
+            },
+          },
+
+          // Apply PostCSS plugins including autoprefixer
+          {
+            loader: 'postcss-loader',
+            options: {
+              config: {
+                path: './tools/postcss.config.js',
+              },
+            },
+          },
+
+          // Compile Less to CSS
+          // https://github.com/webpack-contrib/less-loader
+          // Install dependencies before uncommenting: yarn add --dev less-loader less
+          // {
+          //   test: /\.less$/,
+          //   loader: 'less-loader',
+          // },
+
+          // Compile Sass to CSS
+          // https://github.com/webpack-contrib/sass-loader
+          // Install dependencies before uncommenting: yarn add --dev sass-loader node-sass
+          // {
+          //   test: /\.(scss|sass)$/,
+          //   loader: 'sass-loader',
+          // },
         ],
       },
+
+      // Rules for images
       {
-        test: /\.scss$/,
-        loaders: [
-          'isomorphic-style-loader',
-          `css-loader?${JSON.stringify({ sourceMap: DEBUG, minimize: !DEBUG })}`,
-          'postcss-loader?pack=sass',
-          'sass-loader',
+        test: reImage,
+        oneOf: [
+          // Inline lightweight images into CSS
+          {
+            issuer: reStyle,
+            oneOf: [
+              // Inline lightweight SVGs as UTF-8 encoded DataUrl string
+              {
+                test: /\.svg$/,
+                loader: 'svg-url-loader',
+                options: {
+                  name: staticAssetName,
+                  limit: 4096, // 4kb
+                },
+              },
+
+              // Inline lightweight images as Base64 encoded DataUrl string
+              {
+                loader: 'url-loader',
+                options: {
+                  name: staticAssetName,
+                  limit: 4096, // 4kb
+                },
+              },
+            ],
+          },
+
+          // Or return public URL to image resource
+          {
+            loader: 'file-loader',
+            options: {
+              name: staticAssetName,
+            },
+          },
         ],
       },
-      {
-        test: /\.json$/,
-        loader: 'json-loader',
-      },
+
+      // Convert plain text into JS module
       {
         test: /\.txt$/,
         loader: 'raw-loader',
       },
+
+      // Convert Markdown into HTML
       {
-        test: /\.(png|jpg|jpeg|gif|svg|woff|woff2)$/,
-        loader: 'url-loader',
-        query: {
-          name: DEBUG ? '[path][name].[ext]?[hash]' : '[hash].[ext]',
-          limit: 10000,
-        },
+        test: /\.md$/,
+        loader: path.resolve(__dirname, './lib/markdown-loader.js'),
       },
+
+      // Return public URL for all assets unless explicitly excluded
+      // DO NOT FORGET to update `exclude` list when you adding a new loader
       {
-        test: /\.(eot|ttf|wav|mp3)$/,
+        exclude: [reScript, reStyle, reImage, /\.json$/, /\.txt$/, /\.md$/],
         loader: 'file-loader',
-        query: {
-          name: DEBUG ? '[path][name].[ext]?[hash]' : '[hash].[ext]',
+        options: {
+          name: staticAssetName,
         },
       },
-      {
-        test: /\.jade$/,
-        loader: 'jade-loader',
-      },
+
+      // Exclude dev modules from production build
+      ...(isDebug
+        ? []
+        : [
+            {
+              test: resolvePath(
+                'node_modules/react-deep-force-update/lib/index.js',
+              ),
+              loader: 'null-loader',
+            },
+          ]),
     ],
   },
 
-  resolve: {
-    root: path.resolve(__dirname, '../src'),
-    modulesDirectories: ['node_modules'],
-    extensions: ['', '.webpack.js', '.web.js', '.js', '.jsx', '.json'],
-  },
+  // Don't attempt to continue if there are any errors.
+  bail: !isDebug,
 
-  cache: DEBUG,
-  debug: DEBUG,
+  cache: isDebug,
 
+  // Specify what bundle information gets displayed
+  // https://webpack.js.org/configuration/stats/
   stats: {
+    cached: isVerbose,
+    cachedAssets: isVerbose,
+    chunks: isVerbose,
+    chunkModules: isVerbose,
     colors: true,
-    reasons: DEBUG,
-    hash: VERBOSE,
-    version: VERBOSE,
+    hash: isVerbose,
+    modules: isVerbose,
+    reasons: isDebug,
     timings: true,
-    chunks: VERBOSE,
-    chunkModules: VERBOSE,
-    cached: VERBOSE,
-    cachedAssets: VERBOSE,
+    version: isVerbose,
   },
 
-  postcss(bundler) {
-    return {
-      default: [
-        // Transfer @import rule by inlining content, e.g. @import 'normalize.css'
-        // https://github.com/postcss/postcss-import
-        require('postcss-import')({ addDependencyTo: bundler }),
-        // W3C variables, e.g. :root { --color: red; } div { background: var(--color); }
-        // https://github.com/postcss/postcss-custom-properties
-        require('postcss-custom-properties')(),
-        // W3C CSS Custom Media Queries, e.g. @custom-media --small-viewport (max-width: 30em);
-        // https://github.com/postcss/postcss-custom-media
-        require('postcss-custom-media')(),
-        // CSS4 Media Queries, e.g. @media screen and (width >= 500px) and (width <= 1200px) { }
-        // https://github.com/postcss/postcss-media-minmax
-        require('postcss-media-minmax')(),
-        // W3C CSS Custom Selectors, e.g. @custom-selector :--heading h1, h2, h3, h4, h5, h6;
-        // https://github.com/postcss/postcss-custom-selectors
-        require('postcss-custom-selectors')(),
-        // W3C calc() function, e.g. div { height: calc(100px - 2em); }
-        // https://github.com/postcss/postcss-calc
-        require('postcss-calc')(),
-        // Allows you to nest one style rule inside another
-        // https://github.com/jonathantneal/postcss-nesting
-        require('postcss-nesting')(),
-        // W3C color() function, e.g. div { background: color(red alpha(90%)); }
-        // https://github.com/postcss/postcss-color-function
-        require('postcss-color-function')(),
-        // Convert CSS shorthand filters to SVG equivalent, e.g. .blur { filter: blur(4px); }
-        // https://github.com/iamvdo/pleeease-filters
-        require('pleeease-filters')(),
-        // Generate pixel fallback for "rem" units, e.g. div { margin: 2.5rem 2px 3em 100%; }
-        // https://github.com/robwierzbowski/node-pixrem
-        require('pixrem')(),
-        // W3C CSS Level4 :matches() pseudo class, e.g. p:matches(:first-child, .special) { }
-        // https://github.com/postcss/postcss-selector-matches
-        require('postcss-selector-matches')(),
-        // Transforms :not() W3C CSS Level 4 pseudo class to :not() CSS Level 3 selectors
-        // https://github.com/postcss/postcss-selector-not
-        require('postcss-selector-not')(),
-        // Add vendor prefixes to CSS rules using values from caniuse.com
-        // https://github.com/postcss/autoprefixer
-        require('autoprefixer')({ browsers: AUTOPREFIXER_BROWSERS }),
-      ],
-      sass: [
-        require('autoprefixer')({ browsers: AUTOPREFIXER_BROWSERS }),
-      ],
-    };
-  },
+  // Choose a developer tool to enhance debugging
+  // https://webpack.js.org/configuration/devtool/#devtool
+  devtool: isDebug ? 'cheap-module-inline-source-map' : 'source-map',
 };
 
 //
 // Configuration for the client-side bundle (client.js)
 // -----------------------------------------------------------------------------
 
-const clientConfig = extend(true, {}, config, {
-  entry: './client.js',
+const clientConfig = {
+  ...config,
 
-  output: {
-    filename: DEBUG ? '[name].js?[chunkhash]' : '[name].[chunkhash].js',
-    chunkFilename: DEBUG ? '[name].[id].js?[chunkhash]' : '[name].[id].[chunkhash].js',
-  },
-
+  name: 'client',
   target: 'web',
 
+  entry: {
+    client: ['@babel/polyfill', './src/client.js'],
+  },
+
   plugins: [
-
     // Define free variables
-    // https://webpack.github.io/docs/list-of-plugins.html#defineplugin
-    new webpack.DefinePlugin({ ...GLOBALS, 'process.env.BROWSER': true }),
-
-    // Emit a file with assets paths
-    // https://github.com/sporto/assets-webpack-plugin#options
-    new AssetsPlugin({
-      path: path.resolve(__dirname, '../build'),
-      filename: 'assets.js',
-      processOutput: x => `module.exports = ${JSON.stringify(x)};`,
+    // https://webpack.js.org/plugins/define-plugin/
+    new webpack.DefinePlugin({
+      'process.env.BROWSER': true,
+      __DEV__: isDebug,
     }),
 
-    // Assign the module and chunk ids by occurrence count
-    // Consistent ordering of modules required if using any hashing ([hash] or [chunkhash])
-    // https://webpack.github.io/docs/list-of-plugins.html#occurrenceorderplugin
-    new webpack.optimize.OccurrenceOrderPlugin(true),
+    // Emit a file with assets paths
+    // https://github.com/webdeveric/webpack-assets-manifest#options
+    new WebpackAssetsManifest({
+      output: `${BUILD_DIR}/asset-manifest.json`,
+      publicPath: true,
+      writeToDisk: true,
+      customize: ({ key, value }) => {
+        // You can prevent adding items to the manifest by returning false.
+        if (key.toLowerCase().endsWith('.map')) return false;
+        return { key, value };
+      },
+      done: (manifest, stats) => {
+        // Write chunk-manifest.json.json
+        const chunkFileName = `${BUILD_DIR}/chunk-manifest.json`;
+        try {
+          const fileFilter = file => !file.endsWith('.map');
+          const addPath = file => manifest.getPublicPath(file);
+          const chunkFiles = stats.compilation.chunkGroups.reduce((acc, c) => {
+            acc[c.name] = [
+              ...(acc[c.name] || []),
+              ...c.chunks.reduce(
+                (files, cc) => [
+                  ...files,
+                  ...cc.files.filter(fileFilter).map(addPath),
+                ],
+                [],
+              ),
+            ];
+            return acc;
+          }, Object.create(null));
+          fs.writeFileSync(chunkFileName, JSON.stringify(chunkFiles, null, 2));
+        } catch (err) {
+          console.error(`ERROR: Cannot write ${chunkFileName}: `, err);
+          if (!isDebug) process.exit(1);
+        }
+      },
+    }),
 
-    ...DEBUG ? [] : [
-
-      // Search for equal or similar files and deduplicate them in the output
-      // https://webpack.github.io/docs/list-of-plugins.html#dedupeplugin
-      new webpack.optimize.DedupePlugin(),
-
-      // Minimize all JavaScript output of chunks
-      // https://github.com/mishoo/UglifyJS2#compressor-options
-      new webpack.optimize.UglifyJsPlugin({
-        compress: {
-          screw_ie8: true, // jscs:ignore requireCamelCaseOrUpperCaseIdentifiers
-          warnings: VERBOSE,
-        },
-      }),
-
-      // A plugin for a more aggressive chunk merging strategy
-      // https://webpack.github.io/docs/list-of-plugins.html#aggressivemergingplugin
-      new webpack.optimize.AggressiveMergingPlugin(),
-    ],
+    ...(isDebug
+      ? []
+      : [
+          // Webpack Bundle Analyzer
+          // https://github.com/th0r/webpack-bundle-analyzer
+          ...(isAnalyze ? [new BundleAnalyzerPlugin()] : []),
+        ]),
   ],
 
-  // Choose a developer tool to enhance debugging
-  // http://webpack.github.io/docs/configuration.html#devtool
-  devtool: DEBUG ? 'cheap-module-eval-source-map' : false,
-});
+  // Move modules that occur in multiple entry chunks to a new entry chunk (the commons chunk).
+  optimization: {
+    splitChunks: {
+      cacheGroups: {
+        commons: {
+          chunks: 'initial',
+          test: /[\\/]node_modules[\\/]/,
+          name: 'vendors',
+        },
+      },
+    },
+  },
+
+  // Some libraries import Node modules but don't use them in the browser.
+  // Tell Webpack to provide empty mocks for them so importing them works.
+  // https://webpack.js.org/configuration/node/
+  // https://github.com/webpack/node-libs-browser/tree/master/mock
+  node: {
+    fs: 'empty',
+    net: 'empty',
+    tls: 'empty',
+  },
+};
 
 //
 // Configuration for the server-side bundle (server.js)
 // -----------------------------------------------------------------------------
 
-const serverConfig = extend(true, {}, config, {
-  entry: './server.js',
+const serverConfig = {
+  ...config,
+
+  name: 'server',
+  target: 'node',
+
+  entry: {
+    server: ['@babel/polyfill', './src/server.js'],
+  },
 
   output: {
-    filename: '../../server.js',
+    ...config.output,
+    path: BUILD_DIR,
+    filename: '[name].js',
+    chunkFilename: 'chunks/[name].js',
     libraryTarget: 'commonjs2',
   },
 
-  target: 'node',
+  // Webpack mutates resolve object, so clone it to avoid issues
+  // https://github.com/webpack/webpack/issues/4817
+  resolve: {
+    ...config.resolve,
+  },
+
+  module: {
+    ...config.module,
+
+    rules: overrideRules(config.module.rules, rule => {
+      // Override babel-preset-env configuration for Node.js
+      if (rule.loader === 'babel-loader') {
+        return {
+          ...rule,
+          options: {
+            ...rule.options,
+            presets: rule.options.presets.map(preset =>
+              preset[0] !== '@babel/preset-env'
+                ? preset
+                : [
+                    '@babel/preset-env',
+                    {
+                      targets: {
+                        node: pkg.engines.node.match(/(\d+\.?)+/)[0],
+                      },
+                      modules: false,
+                      useBuiltIns: false,
+                      debug: false,
+                    },
+                  ],
+            ),
+          },
+        };
+      }
+
+      // Override paths to static assets
+      if (
+        rule.loader === 'file-loader' ||
+        rule.loader === 'url-loader' ||
+        rule.loader === 'svg-url-loader'
+      ) {
+        return {
+          ...rule,
+          options: {
+            ...rule.options,
+            emitFile: false,
+          },
+        };
+      }
+
+      return rule;
+    }),
+  },
 
   externals: [
-    /^\.\/assets$/,
-    function filter(context, request, cb) {
-      const isExternal =
-        request.match(/^[@a-z][a-z\/\.\-0-9]*$/i) &&
-        !request.match(/^react-routing/) &&
-        !context.match(/[\\/]react-routing/);
-      cb(null, Boolean(isExternal));
-    },
+    './chunk-manifest.json',
+    './asset-manifest.json',
+    nodeExternals({
+      whitelist: [reStyle, reImage],
+    }),
   ],
 
   plugins: [
-
     // Define free variables
-    // https://webpack.github.io/docs/list-of-plugins.html#defineplugin
-    new webpack.DefinePlugin({ ...GLOBALS, 'process.env.BROWSER': false }),
+    // https://webpack.js.org/plugins/define-plugin/
+    new webpack.DefinePlugin({
+      'process.env.BROWSER': false,
+      __DEV__: isDebug,
+    }),
 
     // Adds a banner to the top of each generated chunk
-    // https://webpack.github.io/docs/list-of-plugins.html#bannerplugin
-    new webpack.BannerPlugin('require("source-map-support").install();',
-      { raw: true, entryOnly: false }),
+    // https://webpack.js.org/plugins/banner-plugin/
+    new webpack.BannerPlugin({
+      banner: 'require("source-map-support").install();',
+      raw: true,
+      entryOnly: false,
+    }),
   ],
 
+  // Do not replace node globals with polyfills
+  // https://webpack.js.org/configuration/node/
   node: {
     console: false,
     global: false,
@@ -302,8 +486,6 @@ const serverConfig = extend(true, {}, config, {
     __filename: false,
     __dirname: false,
   },
-
-  devtool: 'source-map',
-});
+};
 
 export default [clientConfig, serverConfig];