+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ### [Unreleased][unreleased]
 
+- Split the `App` component into `App` setting context variables and `Layout` setting general look and feel of the app (BREAKING CHANGE)
+- Upgrade `history` npm module to v4.x, update `Link` component (BREAKING CHANGE)
+- Remove `core/createHistory.js` in favor of initializing a new history instance inside `server.js` and `client.js` (BREAKING CHANGE)
+- Remove Jade dependency in favor of React-based templates: `src/views/index.jade => src/components/Html`
+  (BREAKING CHANGE) [#711](https://github.com/kriasoft/react-starter-kit/pull/711)
 - Update `isomorphic-style-loader` to `v1.0.0`, it adds comparability with ES2015+ decorators.
   Code such as `export default withStyles(MyComponent, style1, style2)` must be replaced with
   `export default withStyles(style1, style2)(MyComponent)` (BREAKING CHANGE).
@@ -22,6 +27,7 @@ All notable changes to this project will be documented in this file.
 - Move page / screen components into the `src/routes` folder along with the routing information for them (BREAKING CHANGE). [6553936](https://github.com/kriasoft/react-starter-kit/commit/6553936e693e24a8ac6178f4962af15e0ea87dfd)
 
 ### [v0.5.1]
+
 > 2016-03-02
 
 - Remove `Html` React component in favor of compiled Jade templates (`src/views`) (BREAKING CHANGE). [e188388](https://github.com/kriasoft/react-starter-kit/commit/e188388f87069cdc7d501b385d6b0e46c98fed60)
@@ -29,6 +35,7 @@ All notable changes to this project will be documented in this file.
 - Add support for Markdown and HTML for static pages. [#469](https://github.com/kriasoft/react-starter-kit/pull/469), [#477](https://github.com/kriasoft/react-starter-kit/pull/477)
 
 ### [v0.5.0]
+
 > 2016-02-27
 
 - Replace RESTful API endpoint (`src/api`) with GraphQL (`src/data`)
@@ -49,9 +56,10 @@ All notable changes to this project will be documented in this file.
 - Add `CHANGELOG.md` file with a list of notable changes to this project
 
 ### [v0.4.1]
+
 > 2015-10-04
 
-- Replace React Hot Loader (depricated) with React Transform
+- Replace React Hot Loader (deprecated) with React Transform
 - Replace `index.html` template with `Html` (shell) React component
 - Update the deployment script (`tools/deploy.js`), add Git-based deployment example
 - Update ESLint and JSCS settings to use AirBnb JavaScript style guide